requirements.txt
amqp==5.0.5
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
aws-secretsmanager-caching==1.1.1.4
billiard==3.6.3.0
black==20.8b1
boto3==1.16.36
botocore==1.19.36
breadability==0.1.20
cachetools==4.2.0
camelcase==0.2
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cloudinary==1.23.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-braces==1.14.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.8.0
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.20.9
django-filter==2.4.0
django-multi-email-field==0.6.2
django-multiselectfield==0.1.12
django-mysql==3.10.0
django-nine==0.2.4
django-oauth-toolkit==1.3.3
django-rest-auth==0.9.5
django-rest-framework-mongoengine==3.4.1
django-templated-mail==1.1.1
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djangorestframework-xml==2.0.0
djoser==2.0.5
dnspython==1.16.0
docopt==0.6.2
docxcompose==1.3.2
drf-yasg==1.20.0
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
eventlet==0.30.2
Faker==5.0.1
flake8==3.8.4
fpdf==1.7.2
geographiclib==1.50
geopy==2.1.0
google-api-core==1.24.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-dialogflow-cx==0.2.0
google-cloud-talent==2.0.0
googleapis-common-protos==1.52.0
greenlet==1.0.0
grpcio==1.34.0
gunicorn==20.0.4
httplib2==0.18.1
icalendar==4.0.7
idna==2.10
inflection==0.5.1
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
kombu==5.0.2
lazy-object-proxy==1.4.3
libcst==0.3.15
lxml==4.6.1
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.22.0
mypy-extensions==0.4.3
mysqlclient==2.0.2
nltk==3.5
numpy==1.19.3
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.8
pandas==1.1.4
pathspec==0.8.1
Pillow==8.0.1
prompt-toolkit==3.0.18
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycountry==20.7.3
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.3
PyJWT==1.7.1
pylint==2.6.0
pymongo==3.11.2
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
python-decouple==3.3
python-docx==0.8.11
python-http-client==3.3.2
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rest-social-auth==4.2.0
rollbar==0.15.1
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
sendgrid==6.7.1
six==1.15.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sqlparse==0.4.1
starkbank-ecdsa==1.1.1
sumy==0.8.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.53.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
XlsxWriter==1.3.7